annotation_tool/requirements.txt
flask
flask-login
nltk
lxml
pyyaml
